move extend up to DataSource
